Allocate and cleanup resources at the end of the world
