+++ src/main/scala/com.gu.gibbons/lambdas/UserDidNotAnswerLambda.scala
@@ -12,17 +12,15 @@ import services.interpreters._
 class UserDidNotAnswerLambda extends GenericLambda {
   import JsonFormats._
 
-  override def go(settings: Settings, dryRun: Boolean) = 
+  override def go(resources: Resources, logger: LoggingInterpreter, settings: Settings, dryRun: Boolean) = {
+    val bonobo = new BonoboInterpreter(settings, logger, resources.dynamo, resources.http)
+    val email = new EmailInterpreter(settings, logger, resources.email)
+    
     for {
-      logger <- LoggingInterpreter.apply
       _ <- logger.info("Hello")
-      _ <- logger.info("Opening up a connection to Bonobo...")
-      bonobo <- BonoboInterpreter(settings, logger)
-      _ <- logger.info("Opening up a connection to SES...")
-      email <- EmailInterpreter(settings, logger)
-      _ <- logger.info("We're all set, starting...")
       userDidNotAnswer = new UserDidNotAnswer(settings, email, bonobo, logger)
       rDel <- userDidNotAnswer.run(dryRun)
       _ <- logger.info("Goodbye")
     } yield rDel.asJson
+  }
 }
